Moving classes in separate package
